requirements.txt
fastapi
uvicorn
firebase-admin
requests==2.32.3
aiohttp==3.11.18
python-dotenv==1.0.1
apscheduler==3.10.4
google-auth==2.39.0
google-api-core==2.24.2
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
grpcio==1.71.0
grpcio-status==1.71.0